use a more efficient way to call javascript callbacks
previously we constructed a string like
"functioname("arg1,"arg2");" and passed that to
eval_print(), a function that would just execute
the strings content.
the disadvantage of that method is that we need
to escape all the strings passed that way to prevent
easy exploiting.

thanks go to tor8 who showed me how to do it
directly.