add support for names list (353) via nameshandler()
nameshandler may be called several times with a number of nicks in
a specific channel. operator names are prefixed with @. voiced
user's names are prefixed with +, etc.

added support for action handlers with a distinct number of dispatch
arguments and IRC arguments (IRC arguments are what the raw IRC
command is composed of, dispatch arguments the arguments sent to the
dispatch function). that makes it possible to handle more numeric
commands in the same way.