add a no-activity timeout preventing bot from getting "stuck"
for some reason, maybe due to a bug in wolfssl, which is used by rocksock's
ssl layer, it can happen that even though the bot got disconnected (happened
in a netsplit situation, running rocksock_peek on the socket, which calls
CyaSSL_pending() and select() on the socket, no error is reported, and the
main loop calling rsirc_process() never detects the exceptional situation.
for that reason we assume if we didn't receive anything from the IRC server
for 5 minutes, that the connection is dead (there should be regular PINGs
every 1-2 mins usually).